feat :: setPresentUga

+++ src/main/java/com/olympus/uga/domain/family/presentation/dto/response/FamilyInfoRes.java
@@ -12,7 +12,7 @@ public class FamilyInfoRes {
 
     private String familyCode;
     private String familyName;
-    private int presentUgaId;
+    private Long presentUgaId;
     private String profileLink;
     private String representativePhoneNum;
     private List<User> memberList;
